% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{adjust_BB_XYcoords}
\alias{adjust_BB_XYcoords}
\title{This function adjusts coordinates of the bounding box for two sets of rasters
corresponding to 10 m and 20 m S2 bands}
\usage{
adjust_BB_XYcoords(BB_XYcoords)
}
\arguments{
\item{BB_XYcoords}{list. includes pixel coordinates for UL, LL, UR, LR corners
of S2Bands_10m and S2Bands_20m}
}
\value{
BB_XYcoords
}
\description{
This function adjusts coordinates of the bounding box for two sets of rasters
corresponding to 10 m and 20 m S2 bands
}
