% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_ENVI_hdr.R
\name{adjust_ENVI_hdr}
\alias{adjust_ENVI_hdr}
\title{This function adjusts information from ENVI header}
\usage{
adjust_ENVI_hdr(dsn, Bands, sensor = "Unknown", Stretch = FALSE)
}
\arguments{
\item{dsn}{character. path where to store the stack}

\item{Bands}{list. should include 'bandname', and if possible 'wavelength'}

\item{sensor}{character. Name of the sensor used to acquire the image}

\item{Stretch}{boolean. Set TRUE to get 10\% stretching at display for reflectance, mentioned in hdr only}
}
\value{
None
}
\description{
This function adjusts information from ENVI header
}
