% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_radiometric_filter.R
\name{apply_radiometric_filter}
\alias{apply_radiometric_filter}
\title{updates vegetation mask based on initial SCL and readiometric thresholds}
\usage{
apply_radiometric_filter(
  S2_items,
  acq,
  baseline,
  iChar,
  raster_dir,
  mask_path = NULL,
  aoiplot,
  fraction_vegetation,
  offset = 1000,
  RadiometricFilter = NULL,
  asset_cloud
)
}
\arguments{
\item{S2_items}{path for spectral bands required}

\item{acq}{date of acquisition}

\item{baseline}{processing baseeline for acquisition to account for offset}

\item{iChar}{plot ID}

\item{raster_dir}{directory where rasters are stored}

\item{mask_path}{path for binary mask}

\item{aoiplot}{sf.}

\item{fraction_vegetation}{minimum fraction of vegetation required to keep the acquisition}

\item{offset}{value of the offset to apply to the S2 bands}

\item{RadiometricFilter}{list. criterions for cloud, shade, non vegetation}

\item{asset_cloud}{character. asset name for cloud mask}
}
\value{
list of collections per plot
}
\description{
updates vegetation mask based on initial SCL and readiometric thresholds
}
