% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_s2_cdse.R
\name{cdse_download}
\alias{cdse_download}
\title{Download Copernicus Dataspace products listed from STAC search.}
\usage{
cdse_download(
  x,
  outdir,
  unzip = TRUE,
  progress = TRUE,
  dry_run = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{x}{rstac::doc_items}

\item{outdir}{character. output dir}

\item{unzip}{logical. If TRUE (default), the downloaded zip files are unzipped.}

\item{progress}{logical. If TRUE (default), a progress bar is displayed.}

\item{dry_run}{logical. If TRUE, the authetication is tried but the download is not performed.}

\item{overwrite}{logical. If TRUE, existing files are overwritten.}
}
\value{
list of downloaded files
}
\description{
Downloads products from Copernicus Dataspace.
Credentials are read from environment variables, so the easiest is to
set them in .Renviron or .Rprofile files in your home directory.
}
\details{
In ~/.Renviron

\if{html}{\out{<div class="sourceCode shell">}}\preformatted{SEN2PROC_CDSE_USERNAME="username"
SEN2PROC_CDSE_PASSWORD="password"
}\if{html}{\out{</div>}}

In ~/.Rprofile

\if{html}{\out{<div class="sourceCode r">}}\preformatted{Sys.setenv(
 SEN2PROC_CDSE_USERNAME="username",
 SEN2PROC_CDSE_PASSWORD="password")
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
# search for L2A products from 2016-01-01 to 2018-01-31
products <- S2_search(
  start = "2016-01-01", end = "2016-01-31",
  tile = "34UCF", level = "L2A"
)

# download the products
cdse_download(products, outdir = "./data/download")
}

}
