% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_existing_mask.R
\name{check_existing_mask}
\alias{check_existing_mask}
\title{downloads S2 reflectance to update the vegetation mask}
\usage{
check_existing_mask(item_collection, band_url, rootname)
}
\arguments{
\item{item_collection}{path for collection to download}

\item{band_url}{list.}

\item{rootname}{character. directory where rasters are stored}
}
\value{
list of collections per plot
}
\description{
downloads S2 reflectance to update the vegetation mask
}
