% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_order_bands.R
\name{check_order_bands}
\alias{check_order_bands}
\title{checks band order}
\usage{
check_order_bands(acq, patterns)
}
\arguments{
\item{acq}{character.}

\item{patterns}{character.}
}
\value{
acq updated acquisitions
}
\description{
checks band order
}
