% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_s2_mission.R
\name{check_s2_mission}
\alias{check_s2_mission}
\title{This function saves reflectance files}
\usage{
check_s2_mission(S2Sat = NULL, tile_S2, dateAcq_S2)
}
\arguments{
\item{S2Sat}{character. Sentinel-2 mission ('2A' or '2B')}

\item{tile_S2}{character. S2 tile name (2 numbers + 3 letters)}

\item{dateAcq_S2}{double. date of acquisition}
}
\value{
s2mission character. name of the S2 mission (2A or 2B)
}
\description{
This function saves reflectance files
}
