% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_S2tilinggrid.R
\name{check_S2tilinggrid}
\alias{check_S2tilinggrid}
\title{check if S2 tiling grid is available}
\usage{
check_S2tilinggrid(path_S2tilinggrid = NULL)
}
\arguments{
\item{path_S2tilinggrid}{character. path for S2 tiling grid}
}
\value{
valid path for S2 tiling grid
}
\description{
check if S2 tiling grid is available
}
