% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_SI_computed.R
\name{check_SI_computed}
\alias{check_SI_computed}
\title{checks if SI already computed}
\usage{
check_SI_computed(dsn_grid, output_SI, SI_list, siteName)
}
\arguments{
\item{dsn_grid}{list.}

\item{output_SI}{character. path for output directory for SI}

\item{SI_list}{character. list of SI}

\item{siteName}{character. site name}
}
\value{
plots list of plots
}
\description{
checks if SI already computed
}
