% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rasters.R
\name{clean_rasters}
\alias{clean_rasters}
\title{clean raster directory from rasters corresponding to plot of interest}
\usage{
clean_rasters(raster_dir, Acqdates, iChar)
}
\arguments{
\item{raster_dir}{character. directory to store collection corrresponding to each plot}

\item{Acqdates}{character. list fo dates of acquisition}

\item{iChar}{character. plot ID}
}
\value{
none
}
\description{
clean raster directory from rasters corresponding to plot of interest
}
