% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_si_from_grid.R
\name{compute_si_from_grid}
\alias{compute_si_from_grid}
\title{compute spectral indices for each cell of a grid defined over a raster}
\usage{
compute_si_from_grid(
  rast_path,
  mask_path,
  plots,
  siteName,
  si_list,
  output_dir,
  overwrite = FALSE,
  spectral_bands = NULL,
  sensor_name = "sentinel-2",
  ReflFactor = 10000
)
}
\arguments{
\item{rast_path}{character.}

\item{mask_path}{character.}

\item{plots}{list. list of plots (sf object)}

\item{siteName}{character.}

\item{si_list}{character.}

\item{output_dir}{character.}

\item{overwrite}{boolean.}

\item{spectral_bands}{numeric.}

\item{sensor_name}{character.}

\item{ReflFactor}{numeric.}
}
\value{
filename_si list of file paths produced
}
\description{
compute spectral indices for each cell of a grid defined over a raster
}
