% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_geom.R
\name{correct_geom}
\alias{correct_geom}
\title{corrects geometry of acquisition}
\usage{
correct_geom(
  S2_rast,
  output_dir,
  aoi,
  acq,
  bands2correct = c("B8A", "B11", "B12")
)
}
\arguments{
\item{S2_rast}{S2 spatRaster}

\item{output_dir}{character.}

\item{aoi}{sf object}

\item{acq}{date of acquisition}

\item{bands2correct}{character. name of bands to correct from geometry}
}
\value{
list of collections per plot
}
\description{
corrects geometry of acquisition
}
