% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_s2_stack.R
\name{correct_s2_stack}
\alias{correct_s2_stack}
\title{produce S2 stack and correct it if needed (BRDF or B2)
These are very experimental corrections}
\usage{
correct_s2_stack(
  s2_items,
  acq,
  raster_dir,
  aoi,
  offset_B2 = FALSE,
  corr_BRF = FALSE,
  bands_to_correct = c("B8A", "B11", "B12")
)
}
\arguments{
\item{s2_items}{path for spectral bands required}

\item{acq}{date of acquisition}

\item{raster_dir}{directory where rasters are stored}

\item{aoi}{sf.}

\item{offset_B2}{boolean. value of the offset to apply to the S2 band B2}

\item{corr_BRF}{boolean.}

\item{bands_to_correct}{character. name of bands to correct from geometry}
}
\value{
list of collections per plot
}
\description{
produce S2 stack and correct it if needed (BRDF or B2)
These are very experimental corrections
}
