% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_from_to.R
\name{datetime_from_to}
\alias{datetime_from_to}
\title{get datetime based on a set of parameters}
\usage{
datetime_from_to(YYYY, MMDD_start, duration)
}
\arguments{
\item{YYYY}{character. Year of acquisition}

\item{MMDD_start}{character. starting Month and Day of acquisition}

\item{duration}{character. duration in month}
}
\value{
list
}
\description{
get datetime based on a set of parameters
}
