% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{dir_size}
\alias{dir_size}
\title{this function aims at computing directory size}
\usage{
dir_size(path, recursive = TRUE)
}
\arguments{
\item{path}{character. path for directory}

\item{recursive}{boolean . set T if recursive}
}
\value{
size_files numeric. size in bytes
\itemize{
\item image stack
\item path for individual band files corresponding to the stack
\item path for vector (reprojected if needed)
}
}
\description{
this function aims at computing directory size
}
