% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cloudmask.R
\name{download_cloudmask}
\alias{download_cloudmask}
\title{downloads SCL from Planetary}
\usage{
download_cloudmask(
  aoi,
  raster_dir,
  collection_info,
  iChar,
  collection = "sentinel-2-l2a",
  asset_names,
  resolution
)
}
\arguments{
\item{aoi}{geometry corresponding to one point}

\item{raster_dir}{directory where rasters are stored}

\item{collection_info}{description of collection to download}

\item{iChar}{plot ID}

\item{collection}{character. name of the collection}

\item{asset_names}{character.}

\item{resolution}{numeric. spatial resolution (10 or 20)}
}
\value{
list of collections per plot
}
\description{
downloads SCL from Planetary
}
