% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_GeomAcq_s2.R
\name{download_GeomAcq_s2}
\alias{download_GeomAcq_s2}
\title{download from CDSE geometry of acquisition for aoi}
\usage{
download_GeomAcq_s2(aoi, geom_dir, collection_GeomAcq_S2, nbCPU = 1)
}
\arguments{
\item{aoi}{simple feature. vector including all S2 tiles intersecting with aoi}

\item{geom_dir}{character. directory where rasters for geometry are saved}

\item{collection_GeomAcq_S2}{list. STAC collection}

\item{nbCPU}{numeric. number of CPUs}
}
\value{
list of collections per plot
}
\description{
download from CDSE geometry of acquisition for aoi
}
