% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_s2.R
\name{download_s2}
\alias{download_s2}
\title{downloads S2 reflectance data}
\usage{
download_s2(
  aoi,
  raster_dir,
  collection_path,
  iChar,
  resolution,
  S2_items = NULL,
  offset = 1000,
  offset_B2 = FALSE,
  stac_info,
  corr_BRF = FALSE,
  resampling = "near",
  site_name = NULL,
  crs_target = NULL,
  writeoutput = TRUE,
  bands_to_correct = c("B8A", "B11", "B12"),
  overwrite = TRUE
)
}
\arguments{
\item{aoi}{geometry corresponding to one point}

\item{raster_dir}{directory where rasters are stored}

\item{collection_path}{path for collection to download}

\item{iChar}{plot ID}

\item{resolution}{numeric. spatial resolution (10 or 20)}

\item{S2_items}{list. S2 items already downloaded}

\item{offset}{numeric. S2 offset value}

\item{offset_B2}{boolean.}

\item{stac_info}{list.}

\item{corr_BRF}{boolean.}

\item{resampling}{character. resampling method for terra::resample}

\item{site_name}{character. name of the study site}

\item{crs_target}{numeric.}

\item{writeoutput}{boolean. should output file be saved?}

\item{bands_to_correct}{character. name of bands to correct from geometry}

\item{overwrite}{boolean.}
}
\value{
list of collections per plot
}
\description{
downloads S2 reflectance data
}
