% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_s2_acq.R
\name{download_s2_acq}
\alias{download_s2_acq}
\title{downloads individual acquisitions}
\usage{
download_s2_acq(
  acq,
  band_url,
  raster_dir,
  plots_bbox,
  iChar,
  asset_names,
  skipExists = T
)
}
\arguments{
\item{acq}{date.}

\item{band_url}{list.}

\item{raster_dir}{character.}

\item{plots_bbox}{sf.}

\item{iChar}{character.}

\item{asset_names}{list.}

\item{skipExists}{boolean.}
}
\value{
out_file
}
\description{
downloads individual acquisitions
}
