% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_s2_collection.R
\name{download_s2_collection}
\alias{download_s2_collection}
\title{downloads a time series defined by item collection}
\usage{
download_s2_collection(
  collection_path,
  aoi,
  iChar,
  raster_dir,
  mask_path = NULL,
  fraction_vegetation = 5,
  stac_info,
  resolution = 10,
  offset = 1000,
  offset_B2 = FALSE,
  corr_BRF = FALSE,
  p = NULL,
  radiometric_filter = NULL,
  overwrite = TRUE,
  site_name = NULL,
  rast_out = TRUE,
  additional_process = NULL,
  crs_target = NULL,
  original_clouds = TRUE,
  argsin = NULL,
  writeoutput = TRUE,
  resampling = "near",
  bands_to_correct = c("B8A", "B11", "B12")
)
}
\arguments{
\item{collection_path}{character. directory where plot collection is stored}

\item{aoi}{geometry corresponding to one point}

\item{iChar}{plot ID}

\item{raster_dir}{directory where rasters are stored}

\item{mask_path}{path for binary mask}

\item{fraction_vegetation}{numeric. minimum fraction vegetation over plot}

\item{stac_info}{list}

\item{resolution}{numeric. spatial resolution (10 or 20)}

\item{offset}{numeric. offset value}

\item{offset_B2}{boolean.}

\item{corr_BRF}{boolean.}

\item{p}{list.}

\item{radiometric_filter}{list. values for radiometric filter shade cloud vegetation}

\item{overwrite}{boolean.}

\item{site_name}{character. name of the study site}

\item{rast_out}{boolean. should S2 SpatRaster be obtained as output?}

\item{additional_process}{additional process to be applied to S2_items once downloaded}

\item{crs_target}{numeric.}

\item{original_clouds}{boolean}

\item{argsin}{list}

\item{writeoutput}{boolean. should output file be saved?}

\item{resampling}{character. resampling method for terra::resample}

\item{bands_to_correct}{character. name of bands to correct from geometry}
}
\value{
list of collections per plot
}
\description{
downloads a time series defined by item collection
}
