% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_s2collection.R
\name{download_s2collection}
\alias{download_s2collection}
\title{downloads a time series defined by item collection}
\usage{
download_s2collection(
  collection_path,
  aoi,
  iChar,
  raster_dir,
  mask_path = NULL,
  fraction_vegetation = 10,
  collection = "sentinel-2-l2a",
  resolution = 10,
  offset = 1000,
  offset_B2 = F,
  corr_BRF = F,
  p = NULL,
  RadiometricFilter = NULL,
  overwrite = T,
  siteName = NULL,
  rast_out = T,
  additional_process = NULL
)
}
\arguments{
\item{collection_path}{character. directory where plot collection is stored}

\item{aoi}{geometry corresponding to one point}

\item{iChar}{plot ID}

\item{raster_dir}{directory where rasters are stored}

\item{mask_path}{path for binary mask}

\item{fraction_vegetation}{numeric. minimum fraction vegetation over plot}

\item{collection}{character. collection targeted with CDSE}

\item{resolution}{numeric. spatial resolution (10 or 20)}

\item{offset}{numeric. offset value}

\item{offset_B2}{boolean.}

\item{corr_BRF}{boolean.}

\item{p}{list.}

\item{RadiometricFilter}{list. values for radiometric filter shade cloud vegetation}

\item{overwrite}{boolean.}

\item{siteName}{character. name of the study site}

\item{rast_out}{boolean. should S2 SpatRaster be obtained as output?}

\item{additional_process}{additional process to be applied to S2_items once downloaded}
}
\value{
list of collections per plot
}
\description{
downloads a time series defined by item collection
}
