% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_incompleteTiles.R
\name{eliminate_incompleteTiles}
\alias{eliminate_incompleteTiles}
\title{keeps only wanted tiles from a collection}
\usage{
eliminate_incompleteTiles(collection_plot, s2_tiles)
}
\arguments{
\item{collection_plot}{STAC item collection}

\item{s2_tiles}{tiles to select}
}
\value{
collection_plot updated STAC item collection
}
\description{
keeps only wanted tiles from a collection
}
