% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_safe.R
\name{extract_from_safe}
\alias{extract_from_safe}
\title{reads S2 data from SAFE data structure, crops if vector file provided
and writes reflectance data as well as cloud mask}
\usage{
extract_from_safe(safe_path, path_aoi = NULL, resolution = 10, output_dir)
}
\arguments{
\item{safe_path}{character. path for unalterated L2a SAFE product}

\item{path_aoi}{character. path for vector file defining aoi included in SAFE}

\item{resolution}{numeric. buffer applied to vector file (in meters)}

\item{output_dir}{character. path for output: reflectance, mask, metadata}
}
\value{
list. path for reflectance, cloud mask and metadata
}
\description{
reads S2 data from SAFE data structure, crops if vector file provided
and writes reflectance data as well as cloud mask
}
