% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_SI_fromSpatRaster.R
\name{fun_SI_fromSpatRaster}
\alias{fun_SI_fromSpatRaster}
\title{compute spectral indices and composite for a given period of time}
\usage{
fun_SI_fromSpatRaster(S2_refl, S2_mask, argsin)
}
\arguments{
\item{S2_refl}{spatRaster.}

\item{S2_mask}{spatRaster.}

\item{argsin}{list of input arguments}
}
\value{
none
}
\description{
compute spectral indices and composite for a given period of time
}
