% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aoi.R
\name{get_aoi}
\alias{get_aoi}
\title{check if S2 tiling grid is available}
\usage{
get_aoi(aoi_path, crs_target = 4326)
}
\arguments{
\item{aoi_path}{character. path for aoi}

\item{crs_target}{numeric.}
}
\value{
valid path for S2 tiling grid
}
\description{
check if S2 tiling grid is available
}
