% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset.R
\name{get_asset}
\alias{get_asset}
\title{get asset from planetary using gdal}
\usage{
get_asset(asset_url, out_file, plots_bbox)
}
\arguments{
\item{asset_url}{character.}

\item{out_file}{character.}

\item{plots_bbox}{sf..}
}
\value{
vsicurl url
}
\description{
get asset from planetary using gdal
}
