% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset_terra.R
\name{get_asset_terra}
\alias{get_asset_terra}
\title{get asset from planetary using gdal}
\usage{
get_asset_terra(item, asset_names, aoi, collection = "sentinel-2-l2a")
}
\arguments{
\item{item}{character.}

\item{asset_names}{character.}

\item{aoi}{sf.}

\item{collection}{character.}
}
\value{
features
}
\description{
get asset from planetary using gdal
}
