% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_B248_filter.R
\name{get_B248_filter}
\alias{get_B248_filter}
\title{downloads S2 reflectance to update the vegetation mask}
\usage{
get_B248_filter(
  raster_dir,
  mask_path,
  collection_path,
  iChar,
  aoi,
  resolution,
  collection = "sentinel-2-l2a",
  fraction_vegetation,
  offset = 1000,
  RadiometricFilter = NULL
)
}
\arguments{
\item{raster_dir}{directory where rasters are stored}

\item{mask_path}{directory where rasters are stored}

\item{collection_path}{path for collection to download}

\item{iChar}{plot ID}

\item{aoi}{geometry corresponding to one point}

\item{resolution}{numeric. spatial resolution (10 or 20)}

\item{collection}{character.}

\item{fraction_vegetation}{numeric. minimum \% of vegetatiomn to have on plot footprint}

\item{offset}{numeric. value of S2 offset}

\item{RadiometricFilter}{list. list of radiometric filters for shade, clouds and vegetation}
}
\value{
list of collections per plot
}
\description{
downloads S2 reflectance to update the vegetation mask
}
