% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cloudmask.R
\name{get_cloudmask}
\alias{get_cloudmask}
\title{downloads cloud mask}
\usage{
get_cloudmask(
  collection_path,
  aoi,
  iChar,
  raster_dir,
  overwrite = F,
  fraction_vegetation = 10,
  stac_info,
  resolution = 10,
  nbCPU = 1,
  siteName = NULL,
  crs_target = NULL
)
}
\arguments{
\item{collection_path}{path for colection previously saved}

\item{aoi}{geometry corresponding to one point}

\item{iChar}{plot ID}

\item{raster_dir}{directory where rasters are stored}

\item{overwrite}{boolean.}

\item{fraction_vegetation}{numeric. minimum fraction vegetation over plot}

\item{stac_info}{list.}

\item{resolution}{numeric. spatial resolution (10 or 20)}

\item{nbCPU}{numeric. number of CPU}

\item{siteName}{character. name of the study site}

\item{crs_target}{numeric.}
}
\value{
list of cloudmasks
}
\description{
downloads cloud mask
}
