% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collection.R
\name{get_collection}
\alias{get_collection}
\title{gets collections for plot network using microsoft planetary}
\usage{
get_collection(
  aoi,
  S2tiles = NULL,
  datetime,
  FileName,
  overwrite = T,
  cloudcover = 100,
  collection = "sentinel-2-l2a",
  stac_url = NULL
)
}
\arguments{
\item{aoi}{simple feature including a unique polygon}

\item{S2tiles}{list. list of S2 tiles including each sf from list_aoi}

\item{datetime}{list. define period 'from' / 'to' day of acquisition}

\item{FileName}{character. filename for collection}

\item{overwrite}{boolean. should collection and S2 data be overwritten?}

\item{cloudcover}{numeric. maximum cloud cover over tile}

\item{collection}{character. collection targeted with CDSE}

\item{stac_url}{character. URL for STAC endpoint}
}
\value{
list of collections per plot
}
\description{
gets collections for plot network using microsoft planetary
}
