% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{get_corners_From_BB}
\alias{get_corners_From_BB}
\title{get corner coordinates from bounding box defined by pixel coordinates}
\usage{
get_corners_From_BB(input_raster, BBcoords)
}
\arguments{
\item{input_raster}{SpatRaster}

\item{BBcoords}{list. includes pixel coordinates (row, col) for Ur, UL, LR &
LL corners}
}
\value{
Corners coordinates
}
\description{
get corner coordinates from bounding box defined by pixel coordinates
}
