% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dateAcq.R
\name{get_dateAcq}
\alias{get_dateAcq}
\title{gets acquisition date from planetary URL / S2 products / item collection}
\usage{
get_dateAcq(
  band_url = NULL,
  S2product = NULL,
  stac_info = NULL,
  item_collection = NULL,
  collection = "sentinel-2-l2a"
)
}
\arguments{
\item{band_url}{character.}

\item{S2product}{character.}

\item{stac_info}{list.}

\item{item_collection}{list.}

\item{collection}{character.}
}
\value{
DateAcq dates of acquisition
}
\description{
gets acquisition date from planetary URL / S2 products / item collection
}
