% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{get_datetime_from_S2prod}
\alias{get_datetime_from_S2prod}
\title{Using the PRODUCT_ID, this function help to recreate correctly start, stop and generation date in the XML
programmed by romain.dous@inrae.fr}
\usage{
get_datetime_from_S2prod(name_time, info_type = "start")
}
\arguments{
\item{name_time}{character. The PRODUCT_ID (or other information if required)}

\item{info_type}{character. start ; stop ; generation depending of the XML line}
}
\value{
date_product character. Correct date for the XML line
}
\description{
Using the PRODUCT_ID, this function help to recreate correctly start, stop and generation date in the XML
programmed by romain.dous@inrae.fr
}
