% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_footprint_from_bbox_aoi.R
\name{get_footprint_from_bbox_aoi}
\alias{get_footprint_from_bbox_aoi}
\title{return S2 tiles intersecting with the footprint of a vector file}
\usage{
get_footprint_from_bbox_aoi(
  aoi_path = NULL,
  bbox = NULL,
  output_dir,
  overwrite = TRUE
)
}
\arguments{
\item{aoi_path}{character. path for vector file defining aoi}

\item{bbox}{bbox}

\item{output_dir}{character. path for output directory}

\item{overwrite}{boolean. if needs overwrite}
}
\value{
list of path corresponding to output files
}
\description{
return S2 tiles intersecting with the footprint of a vector file
}
