% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GeomAcq_s2_coll.R
\name{get_GeomAcq_s2_coll}
\alias{get_GeomAcq_s2_coll}
\title{get collection for geometry of acquisition corresponding to the aoi}
\usage{
get_GeomAcq_s2_coll(
  aoi,
  datetime,
  collection = "sentinel-2-l2a",
  OAuthToken,
  cloudcover,
  output_dir,
  overwrite = F
)
}
\arguments{
\item{aoi}{spatial feature. vector including all S2 tiles intersecting with aoi}

\item{datetime}{list. define period 'from' / 'to' day of acquisition}

\item{collection}{character. collection targeted with CDSE}

\item{OAuthToken}{authentication token for CDSE}

\item{cloudcover}{numeric. maximum cloud cover over tile}

\item{output_dir}{character. main output directory}

\item{overwrite}{boolean. should collection and S2 data be overwritten?}
}
\value{
list of collections per plot
}
\description{
get collection for geometry of acquisition corresponding to the aoi
}
