% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_aoi.R
\name{get_grid_aoi}
\alias{get_grid_aoi}
\title{return vector grid corresponding to an area of interest}
\usage{
get_grid_aoi(aoi_path, cellsize = 10000, output_dir, crs_target = NULL)
}
\arguments{
\item{aoi_path}{character. path for vector file defining aoi}

\item{cellsize}{numeric. square cell size in meters}

\item{output_dir}{character. output path}

\item{crs_target}{crs to be applied to data collected}
}
\value{
dsn_grid path for grid of aoi
}
\description{
return vector grid corresponding to an area of interest
}
