% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mainmask.R
\name{get_mainmask}
\alias{get_mainmask}
\title{extract info from main mask corresponding to aoi}
\usage{
get_mainmask(mask_path, S2_dl, aoiplot)
}
\arguments{
\item{mask_path}{path for main mask defined over full region}

\item{S2_dl}{plot ID}

\item{aoiplot}{geometry corresponding to aoi}
}
\value{
list of collections per plot
}
\description{
extract info from main mask corresponding to aoi
}
