% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_plots.R
\name{get_missing_plots}
\alias{get_missing_plots}
\title{return vector grid corresponding to an area of interest}
\usage{
get_missing_plots(plots, pursue_existing, datetime, output_dir)
}
\arguments{
\item{plots}{character. path for vector file defining aoi}

\item{pursue_existing}{numeric. square cell size in meters}

\item{datetime}{character. output path}

\item{output_dir}{character. path corresponding to output dir}
}
\value{
missing plot ID
}
\description{
return vector grid corresponding to an area of interest
}
