% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OAuth_client.R
\name{get_OAuth_client}
\alias{get_OAuth_client}
\title{get id and pwd from CDSE OAuth client
https://shapps.dataspace.copernicus.eu/dashboard/#/account/settings}
\usage{
get_OAuth_client()
}
\value{
OAuth_client list of id and pwd for CDSE OAuth client
}
\description{
get id and pwd from CDSE OAuth client
https://shapps.dataspace.copernicus.eu/dashboard/#/account/settings
}
