% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_tile.R
\name{get_plot_tile}
\alias{get_plot_tile}
\title{identifies S2 tile fullly including a plot}
\usage{
get_plot_tile(aoi, S2_footprint)
}
\arguments{
\item{aoi}{sf object}

\item{S2_footprint}{sf object}
}
\value{
buffer size
}
\description{
identifies S2 tile fullly including a plot
}
