% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quicklook.R
\name{get_quicklook}
\alias{get_quicklook}
\title{updates vegetation mask based on initial SCL and readiometric thresholds}
\usage{
get_quicklook(
  raster_path,
  wavelength = NULL,
  driver = "png",
  fact = NULL,
  normalize = T
)
}
\arguments{
\item{raster_path}{character. path for raster}

\item{wavelength}{numeric. wavelengths corresponding to spectral bands. ENVI HDR expected if NULL}

\item{driver}{character. driver / extension for raster}

\item{fact}{numeric. aggregation factor}

\item{normalize}{boolean. should normalization be applied?}
}
\value{
none
}
\description{
updates vegetation mask based on initial SCL and readiometric thresholds
}
