% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s2_angles.R
\name{get_s2_angles}
\alias{get_s2_angles}
\title{get S2 minimum and maximum angles for geometry of acquisition}
\usage{
get_s2_angles(path_angles, path_bbox, dateAcq = NULL)
}
\arguments{
\item{path_angles}{character path for raster angles of acquisition dowloaded with preprocS2}

\item{path_bbox}{character path for GPKG defining bounding box of study site}

\item{dateAcq}{date. date of aquisition}
}
\value{
list
}
\description{
get S2 minimum and maximum angles for geometry of acquisition
}
