% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{get_S2_bands}
\alias{get_S2_bands}
\title{This function returns path for the spectral bands to be used}
\usage{
get_S2_bands(Path_dir_S2, S2source = "SAFE", resolution = 10)
}
\arguments{
\item{Path_dir_S2}{character. Path for the directory containing S2 data. either L2A .SAFE S2 file or THEIA directory}

\item{S2source}{character. defines if data comes from SciHub as SAFE directory, from THEIA or from LaSRC}

\item{resolution}{numeric. spatial resolution of the final image: 10m or 20m}
}
\value{
ListBands list. contains path for spectral bands corresponding to 10m and 20m resolution
}
\description{
This function returns path for the spectral bands to be used
}
