% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{get_S2_bands_from_LaSRC}
\alias{get_S2_bands_from_LaSRC}
\title{This function returns path for the spectral bands in LaSRC directory}
\usage{
get_S2_bands_from_LaSRC(Path_dir_S2, resolution = 10)
}
\arguments{
\item{Path_dir_S2}{character. Path for the SAFE directory containing S2 data}

\item{resolution}{numeric. spatial resolution of the final image: 10m or 20m}
}
\value{
ListBands list. contains path for spectral bands corresponding to 10m and 20m resolution, as well name of as granule
}
\description{
This function returns path for the spectral bands in LaSRC directory
}
