% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{get_S2_bands_from_RAW}
\alias{get_S2_bands_from_RAW}
\title{This function returns path for the spectral bands in S2 directory}
\usage{
get_S2_bands_from_RAW(Path_dir_S2, Path_dir_Mask = NULL)
}
\arguments{
\item{Path_dir_S2}{character. Path for the SAFE directory containing S2 data}

\item{Path_dir_Mask}{character. path for S2 mask directory}
}
\value{
ListBands list. contains path for spectral bands corresponding to 10m and 20m resolution, as well name of as granule
}
\description{
This function returns path for the spectral bands in S2 directory
}
