% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{get_S2_bands_from_RGEE}
\alias{get_S2_bands_from_RGEE}
\title{This function returns path for the spectral bands in GEE directory
programmed by romain.dous@inrae.fr}
\usage{
get_S2_bands_from_RGEE(Path_dir_S2)
}
\arguments{
\item{Path_dir_S2}{character. Path for the SAFE directory containing S2 data}
}
\value{
ListBands list. contains path for spectral bands corresponding to 10m and 20m resolution, as well name of as granule
}
\description{
This function returns path for the spectral bands in GEE directory
programmed by romain.dous@inrae.fr
}
