% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_S2_bands_from_THEIA.R
\name{get_s2_bands_from_THEIA}
\alias{get_s2_bands_from_THEIA}
\title{This function returns path for the spectral bands in THEIA directory}
\usage{
get_s2_bands_from_THEIA(Path_dir_S2, resolution = 10, fre_sre = "FRE")
}
\arguments{
\item{Path_dir_S2}{character. Path for the SAFE directory containing S2 data}

\item{resolution}{numeric. spatial resolution of the final image: 10m or 20m}

\item{fre_sre}{character. SRE or FRE products from THEIA}
}
\value{
ListBands list. contains path for spectral bands corresponding to 10m and 20m resolution, as well name of as granule
}
\description{
This function returns path for the spectral bands in THEIA directory
}
