% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s2_baseline.R
\name{get_s2_baseline}
\alias{get_s2_baseline}
\title{gets processing baseline from S2 SAFE product}
\usage{
get_s2_baseline(stac_info, item_collection)
}
\arguments{
\item{stac_info}{list.}

\item{item_collection}{path for collection to download}
}
\value{
baseline
}
\description{
gets processing baseline from S2 SAFE product
}
