% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s2_collection.R
\name{get_s2_collection}
\alias{get_s2_collection}
\title{downloads Sentinel-2 time series for a vectorized plot network}
\usage{
get_s2_collection(
  plots,
  s2_tiles = NULL,
  datetime,
  output_dir,
  mask_path = NULL,
  cloudcover = 100,
  fraction_vegetation = 5,
  resolution = 10,
  stac_info,
  overwrite = FALSE,
  nbCPU = 1,
  overwrite_collection = TRUE,
  offset = 1000,
  offset_B2 = FALSE,
  corr_BRF = FALSE,
  radiometric_filter = NULL,
  site_name = NULL,
  rast_out = TRUE,
  additional_process = NULL,
  crs_target = NULL,
  original_clouds = TRUE,
  argsin = NULL,
  writeoutput = TRUE,
  resampling = "near",
  bands_to_correct = c("B8A", "B11", "B12")
)
}
\arguments{
\item{plots}{list. sf objects corresponding to polygons}

\item{s2_tiles}{list. S2 tiles corresponding to the plots}

\item{datetime}{list. define period 'from' / 'to' day of acquisition}

\item{output_dir}{character. output directory}

\item{mask_path}{character. path for mask}

\item{cloudcover}{numeric. maximum cloud cover over tile}

\item{fraction_vegetation}{numeric. minimum fraction vegetation to consider acquisition}

\item{resolution}{numeric. spatial resolution (10 or 20)}

\item{stac_info}{list. stac provider, name of collection url for stac}

\item{overwrite}{boolean. should collection and S2 data be overwritten?}

\item{nbCPU}{numeric. number of threads to work with}

\item{overwrite_collection}{boolean.}

\item{offset}{numeric. S2 reflectance offset}

\item{offset_B2}{boolean. should an offset be applied to normalize B2?}

\item{corr_BRF}{boolean.}

\item{radiometric_filter}{list.}

\item{site_name}{character. name of the study site}

\item{rast_out}{boolean. should S2 SpatRaster be obtained as output?}

\item{additional_process}{additional process to be applied to S2_items once downloaded}

\item{crs_target}{numeric.}

\item{original_clouds}{should original cloud mask be used or not?}

\item{argsin}{list}

\item{writeoutput}{boolean. should output file be saved?}

\item{resampling}{character. resampling method for terra::resample}

\item{bands_to_correct}{character. name of bands to correct from geometry}
}
\value{
list of collections per plot
}
\description{
downloads Sentinel-2 time series for a vectorized plot network
}
