% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s2_footprint.R
\name{get_s2_footprint}
\alias{get_s2_footprint}
\title{return footprint corresponding to an vector file}
\usage{
get_s2_footprint(dsn)
}
\arguments{
\item{dsn}{character path for vector file}
}
\value{
footprint
}
\description{
return footprint corresponding to an vector file
}
