% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{get_S2_L1C_Image}
\alias{get_S2_L1C_Image}
\title{download S2 L1C data from Copernicus hub or Google Cloud}
\usage{
get_S2_L1C_Image(
  list_safe,
  l1c_path,
  path_vector,
  time_interval,
  GoogleCloud = FALSE,
  ForceGoogle = FALSE
)
}
\arguments{
\item{list_safe}{safe object. produced with sen2r::s2_list}

\item{l1c_path}{character. path for storage of L1C image}

\item{path_vector}{path for a vector file}

\item{time_interval}{dates. time interval for S2 query}

\item{GoogleCloud}{boolean. set to TRUE if google cloud SDK is installed and}

\item{ForceGoogle}{boolean. set to TRUE if only google requested
sen2r configured as an alternative hub for S2 download}
}
\value{
prodName character. S2 Product name
}
\description{
download S2 L1C data from Copernicus hub or Google Cloud
}
