% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{get_S2_L2A_Image_from_GEE}
\alias{get_S2_L2A_Image_from_GEE}
\title{This function downloads S2 data from GEE based on a vector footprint}
\usage{
get_S2_L2A_Image_from_GEE(
  l2a_path,
  path_vector,
  dateAcq,
  dateEnd = NULL,
  email,
  cloud_cover = 10,
  ImageCollection = "COPERNICUS/S2_SR",
  B10m = c("B2", "B3", "B4", "B8"),
  B20m = c("B5", "B6", "B7", "B8A", "B11", "B12"),
  B_cloud = c("MSK_CLDPRB")
)
}
\arguments{
\item{l2a_path}{character. Storage option -> Path where download S2 datas}

\item{path_vector}{character. Path for vector file}

\item{dateAcq}{character. Only or starting acquisition date}

\item{dateEnd}{character. Ending acquisition date for multiple acquisition}

\item{email}{character. Personal mail with RGEE autorisations (datas will be download on the Google Drive of thie account)}

\item{cloud_cover}{numeric. Percent of clouds tolerate on each acquisition.}

\item{ImageCollection}{character. image data collection}

\item{B10m}{character. vector containing the name of the 10m bands to be downloaded}

\item{B20m}{character. vector containing the name of the 20m bands to be downloaded}

\item{B_cloud}{character. vector containing the name of the Cloud band to be downloaded}
}
\value{
character. path corresponding to acquisitions
}
\description{
This function downloads S2 data from GEE based on a vector footprint
}
