% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{get_S2_level}
\alias{get_S2_level}
\title{This function check S2 data level:
\itemize{
\item L2A: already atmospherically corrected
\item L1C: requires atmospheric corrections with sen2cor
}}
\usage{
get_S2_level(prodName)
}
\arguments{
\item{prodName}{character. original name for the S2 image}
}
\value{
S2Level character. S2 level: L1C or L2A
}
\description{
This function check S2 data level:
\itemize{
\item L2A: already atmospherically corrected
\item L1C: requires atmospheric corrections with sen2cor
}
}
