% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_S2_offset.R
\name{get_s2_offset}
\alias{get_s2_offset}
\title{This function gets offset from S2 images produced from sen2cor or LaSRC}
\usage{
get_s2_offset(MTD_MSI = NULL, MTD_LaSRC = NULL)
}
\arguments{
\item{MTD_MSI}{character. path for MTD_MSI file}

\item{MTD_LaSRC}{character. path for MTD_LaSRC file (specific to LaSRC)}
}
\value{
S2_offset list. info required to correct offset
}
\description{
This function gets offset from S2 images produced from sen2cor or LaSRC
}
