% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s2_raster.R
\name{get_s2_raster}
\alias{get_s2_raster}
\title{return S2 tiles intersecting with the footprint of a vector file}
\usage{
get_s2_raster(
  aoi_path = NULL,
  bbox = NULL,
  mask_path = NULL,
  datetime,
  output_dir,
  site_name = NULL,
  stac_info = NULL,
  options = NULL
)
}
\arguments{
\item{aoi_path}{character. path for vector file defining aoi}

\item{bbox}{bbox}

\item{mask_path}{character. path for a vector or raster mask}

\item{datetime}{character or date or list defining date of acquisition or time range}

\item{output_dir}{character. path for output directory}

\item{site_name}{character. name of the study site}

\item{stac_info}{list. stac provider, name of collection url for stac}

\item{options}{may include
\itemize{
\item cloudcover numeric. cloud cover
\item path_S2_tiling_grid character. path for the Sentinel-2_tiling_grid.kml file
\item overwrite boolean.
\item geom_acq boolean.
\item collection character.
\item stac_url character.
\item additional_process additional process to be applied to S2_items once downloaded
\item bands_to_correct character. name of bands to correct from geometry
\item fraction_vegetation numeric. minimum fraction vegetation to consider acquisition
\item radiometric_filter list.
}}
}
\value{
list of path corresponding to output files
}
\description{
return S2 tiles intersecting with the footprint of a vector file
}
