% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{get_safe}
\alias{get_safe}
\title{get safe name based on granule path}
\usage{
get_safe(granule_path)
}
\arguments{
\item{granule_path}{character. path for granule in SAFE file}
}
\value{
safe name of SAFE file
}
\description{
get safe name based on granule path
}
