% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_si_tiles_from_raster.R
\name{get_si_tiles_from_raster}
\alias{get_si_tiles_from_raster}
\title{computes spectral index corresponding to a vector footprint from a raster
the raster is expecte to be sentinel-2 or landsat}
\usage{
get_si_tiles_from_raster(
  aoi,
  aoi_ID,
  rast_path,
  mask_path = NULL,
  si_list,
  output_dir,
  overwrite = FALSE,
  spectral_bands,
  sensor_name = "sentinel-2",
  site_name,
  ReflFactor = 10000,
  p = NULL
)
}
\arguments{
\item{aoi}{spatVector.}

\item{aoi_ID}{character.}

\item{rast_path}{character.}

\item{mask_path}{character.}

\item{si_list}{character.}

\item{output_dir}{character.}

\item{overwrite}{boolean.}

\item{spectral_bands}{numeric.}

\item{sensor_name}{character. either sentinel-2 or landsat}

\item{site_name}{character.}

\item{ReflFactor}{numeric.}

\item{p}{object}
}
\value{
filename_si list of file paths produced
}
\description{
computes spectral index corresponding to a vector footprint from a raster
the raster is expecte to be sentinel-2 or landsat
}
