% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stac_info.R
\name{get_stac_info}
\alias{get_stac_info}
\title{define stac provider and collection based on a list of information provided
as input}
\usage{
get_stac_info(stac_info = NULL)
}
\arguments{
\item{stac_info}{list. contains provider, stac_url, and collection}
}
\value{
list of collections per plot
}
\description{
define stac provider and collection based on a list of information provided
as input
}
