% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{get_tile}
\alias{get_tile}
\title{This function gets tile from S2 image}
\usage{
get_tile(prodName)
}
\arguments{
\item{prodName}{character. original name for the S2 image}
}
\value{
TileName character
}
\description{
This function gets tile from S2 image
}
