% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetImages_par.R
\name{GetImages_par}
\alias{GetImages_par}
\title{function to call in parallel to get images from CDSE}
\usage{
GetImages_par(
  fileName,
  time_range,
  bbox,
  collection,
  script,
  format,
  mosaicking_order,
  resolution,
  buffer,
  mask,
  token,
  p = NULL
)
}
\arguments{
\item{fileName}{character.}

\item{time_range}{list. define period 'from' / 'to' day of acquisition}

\item{bbox}{numeric. maximum cloud cover over tile}

\item{collection}{character. collection targeted with CDSE}

\item{script}{character. path for script}

\item{format}{character.}

\item{mosaicking_order}{character.}

\item{resolution}{numeric.}

\item{buffer}{numeric.}

\item{mask}{boolean.}

\item{token}{list.}

\item{p}{list.}
}
\value{
rast_out
}
\description{
function to call in parallel to get images from CDSE
}
