% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{granuleIdentifier_creation}
\alias{granuleIdentifier_creation}
\title{This function was made to correctly create the granule line corresponding to a RGEE data acquisition
programmed by romain.dous@inrae.fr}
\usage{
granuleIdentifier_creation(datastrip_id, granule_id)
}
\arguments{
\item{datastrip_id}{character. properties$DATASTRIP_ID}

\item{granule_id}{character. properties$GRANULE_ID}
}
\value{
granule character. GRANULE
}
\description{
This function was made to correctly create the granule line corresponding to a RGEE data acquisition
programmed by romain.dous@inrae.fr
}
