% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{identification_Acq_features}
\alias{identification_Acq_features}
\title{This function is used to find the good features data for an acquisition
programmed by romain.dous@inrae.fr}
\usage{
identification_Acq_features(Acq, feat)
}
\arguments{
\item{Acq}{character. Information of the acquisition. example : "20220321T104731_20220321T105204_T31UFQ".}

\item{feat}{list. Total of features information for all acquisitions. Equivalent of features_MTD$features.}
}
\value{
features_Acq list. Features information only for the targeted acquisition.
stop if the ACq is not present in the features data.
}
\description{
This function is used to find the good features data for an acquisition
programmed by romain.dous@inrae.fr
}
