% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets.R
\name{list_assets}
\alias{list_assets}
\title{checks if assets from item list already exist}
\usage{
list_assets(S2_items, asset_names, dateAcqs)
}
\arguments{
\item{S2_items}{S2 items}

\item{asset_names}{character.}

\item{dateAcqs}{dates.}
}
\value{
list of collections per plot
}
\description{
checks if assets from item list already exist
}
