% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cloud_dl.R
\name{list_cloud_dl}
\alias{list_cloud_dl}
\title{checks if cloud data already downloaded}
\usage{
list_cloud_dl(raster_dir, days, iChar, asset_names)
}
\arguments{
\item{raster_dir}{character. directory to store collection corrresponding to each plot}

\item{days}{character. list fo dates of acquisition}

\item{iChar}{character. plot ID}

\item{asset_names}{character.}
}
\value{
list of collections per plot
}
\description{
checks if cloud data already downloaded
}
