% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_SCL_dl.R
\name{list_SCL_dl}
\alias{list_SCL_dl}
\title{checks if SCL data already downloaded}
\usage{
list_SCL_dl(raster_dir, days, iChar)
}
\arguments{
\item{raster_dir}{character. directory to store collection corrresponding to each plot}

\item{days}{character. list fo dates of acquisition}

\item{iChar}{character. plot ID}
}
\value{
list of collections per plot
}
\description{
checks if SCL data already downloaded
}
