% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vsicurl_theia_url.R
\name{make_vsicurl_theia_url}
\alias{make_vsicurl_theia_url}
\title{converts URL TO vsicurl_url for THEIA data}
\usage{
make_vsicurl_theia_url(base_url)
}
\arguments{
\item{base_url}{character.}
}
\value{
vsicurl url
}
\description{
converts URL TO vsicurl_url for THEIA data
}
