% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vsicurl_url.R
\name{make_vsicurl_url}
\alias{make_vsicurl_url}
\title{converts URL TO vsicurl_url}
\usage{
make_vsicurl_url(base_url, collection = "sentinel-2-l2a")
}
\arguments{
\item{base_url}{character.}

\item{collection}{character.}
}
\value{
vsicurl url
}
\description{
converts URL TO vsicurl_url
}
