% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocS2-package.R
\docType{package}
\name{preprocS2-package}
\alias{preprocS2}
\alias{preprocS2-package}
\title{preprocS2: preprocessing of Sentinel-2 data downloaded from various data hubs / produced from various atmospheric correction methods}
\description{
The package preprocS2 allows production of raster stack from Sentinel-2 directories, as well as cropping, production of cloud mask....
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/jbferet/preprocS2}
  \item Report bugs at \url{https://gitlab.com/jbferet/preprocS2/issues}
}

}
\author{
\strong{Maintainer}: Jean-Baptiste Feret \email{jb.feret@teledetection.fr} (\href{https://orcid.org/0000-0002-0151-1334}{ORCID}) [copyright holder]

}
\keyword{internal}
