% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_geom_acq_tile.R
\name{produce_geom_acq_tile}
\alias{produce_geom_acq_tile}
\title{extracts geometry of acquisition for each plot}
\usage{
produce_geom_acq_tile(list_aoi, collection_dir_geom, geom_dir, nbCPU)
}
\arguments{
\item{list_aoi}{path for spectral bands required}

\item{collection_dir_geom}{collection.}

\item{geom_dir}{character.}

\item{nbCPU}{numeric.}
}
\value{
list of collections per plot
}
\description{
extracts geometry of acquisition for each plot
}
