% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mask_path_terra.R
\name{read_mask_path_terra}
\alias{read_mask_path_terra}
\title{extract info from main mask corresponding to aoi}
\usage{
read_mask_path_terra(mask_path)
}
\arguments{
\item{mask_path}{path for main mask defined over full region}
}
\value{
spatRaster or spatVector
}
\description{
extract info from main mask corresponding to aoi
}
