% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{read_raster}
\alias{read_raster}
\title{This function reads a raster stack, and gets footprint as pixel coordinates or vector file as input}
\usage{
read_raster(path_raster, path_vector = NULL, BBpix = NULL)
}
\arguments{
\item{path_raster}{character. path for raster file}

\item{path_vector}{character. path for vector file}

\item{BBpix}{list. coordinates of pixels corresponding to a bounding box}
}
\value{
starsobj stars object corresponding to raster or raster subset
}
\description{
This function reads a raster stack, and gets footprint as pixel coordinates or vector file as input
}
