% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_s2_bands.R
\name{read_s2_bands}
\alias{read_s2_bands}
\title{This function reads a list of files corresponding to S2 bands
S2 bands are expected to have uniform spatial resolution and footprint}
\usage{
read_s2_bands(
  S2_Bands,
  aoi = NULL,
  resolution = 10,
  interpolation = "nearest_neighbour"
)
}
\arguments{
\item{S2_Bands}{list. list of S2 bands obtained from get_s2_bands}

\item{aoi}{sf object.}

\item{resolution}{numeric. resolution of final product}

\item{interpolation}{character. method for resampling. default = 'nearest_neighbour'}
}
\value{
Stack_S2 list. contains stack of S2 bands
}
\description{
This function reads a list of files corresponding to S2 bands
S2 bands are expected to have uniform spatial resolution and footprint
}
