% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{read_S2bands}
\alias{read_S2bands}
\title{This function reads a list of files corresponding to S2 bands
S2 bands are expected to have uniform spatial resolution and footprint}
\usage{
read_S2bands(
  S2_Bands,
  path_vector = NULL,
  resolution = 10,
  interpolation = "nearest_neighbour"
)
}
\arguments{
\item{S2_Bands}{list. list of S2 bands obtained from get_S2_bands}

\item{path_vector}{path for a vector file}

\item{resolution}{numeric. resolution of final product}

\item{interpolation}{character. method for resampling. default = 'nearest_neighbour'}
}
\value{
Stack_S2 list. contains stack of S2 bands
}
\description{
This function reads a list of files corresponding to S2 bands
S2 bands are expected to have uniform spatial resolution and footprint
}
