% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{recup_name_crs}
\alias{recup_name_crs}
\title{With the ESPG, this function writes the CRS name corresponding
programmed by romain.dous@inrae.fr}
\usage{
recup_name_crs(crs_init)
}
\arguments{
\item{crs_init}{character. EPSG}
}
\value{
name_crs character. Complete name of the ESPG
}
\description{
With the ESPG, this function writes the CRS name corresponding
programmed by romain.dous@inrae.fr
}
