% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproject_shp.R
\name{reproject_shp}
\alias{reproject_shp}
\title{This function reprojects a shapefile and saves reprojected shapefile}
\usage{
reproject_shp(path_vector_init, SpatialObj, path_vector_reproj)
}
\arguments{
\item{path_vector_init}{character. path for a shapefile to be reprojected}

\item{SpatialObj}{object. spatRaster or spatVector}

\item{path_vector_reproj}{character. path for the reprojected shapefile}
}
\value{
path_vector character. path of the shapefile
\itemize{
\item path_vector_init if the vector did not need reprojection
\item path_vector_reproj if the vector needed reprojection
}
}
\description{
This function reprojects a shapefile and saves reprojected shapefile
}
