% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{S2_band_StandardizeName}
\alias{S2_band_StandardizeName}
\title{This function standardizes S2 band names downloaded with GEE (from BX to B0X)
programmed by romain.dous@inrae.fr}
\usage{
S2_band_StandardizeName(band)
}
\arguments{
\item{band}{character. Original band name (e.g. B2)}
}
\value{
true_num character. standardized band name (e.g. B02)
}
\description{
This function standardizes S2 band names downloaded with GEE (from BX to B0X)
programmed by romain.dous@inrae.fr
}
