% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{S2_from_L1C_to_L2A}
\alias{S2_from_L1C_to_L2A}
\title{perform atmospheric corrections to convert L1C to L2A data with Sen2cor}
\usage{
S2_from_L1C_to_L2A(prodName, l1c_path, l2a_path, datePattern, tmp_path = NULL)
}
\arguments{
\item{prodName}{character. produced with sen2r::s2_list}

\item{l1c_path}{character. path of directory where L1C image is stored}

\item{l2a_path}{character. path of directory where L2A image is stored}

\item{datePattern}{character. pattern corresponding to date of acquisition to identify L2A directory}

\item{tmp_path}{character. path of temporary directory where L2A image is stored
sen2r configured as an alternative hub for S2 download}
}
\value{
PathL2A character. S2 Product name
}
\description{
perform atmospheric corrections to convert L1C to L2A data with Sen2cor
}
