% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_s2_cdse.R
\name{s2_search}
\alias{s2_search}
\title{Search helper for Copernicus DataSpace}
\usage{
s2_search(start, end = NULL, tile = NULL, level = "L2A", id_pat = NULL)
}
\arguments{
\item{start}{start date "YYYY-MM-DD" or Date class.}

\item{end}{end date "YYYY-MM-DD" or Date class. If NULL, end = start + 1 day}

\item{tile}{MGRS tile, e.g. "34UCF"}

\item{level}{L1C or L2A}

\item{id_pat}{pattern to look for in product id, see details}
}
\value{
rstac::doc_items
}
\description{
This helper is wrapper of rstac with
the most used filters, to make it easier to search.
For more options, use directly rstac.
and \code{\link[rstac:stac]{rstac::stac()}}.
}
\details{
The \code{id_pat} argument is a CQL2 pattern, sued with \code{id \%like\% {{id_pat}}}.
Character \verb{\%} is a wild card for string
Character \verb{_} is a wild card for single character
Character \verb{\\\\} is the escape
Thus, the argument "\%\\\emph{N0500\\}\%" is lokking for any id containing string "\emph{N0500}"
}
\examples{
\dontrun{
# search for L2A products from 2016-01-01 to 2018-01-31 with
# - tile: 34UCF
# - processing baseline: N0500

products <- s2_search(
  start = "2016-01-01",
  end = "2018-01-31",
  tile = "34UCF",
  level = "L2A",
  id_pat = "\%\\\\_N0500\\\\_\%"
)

# an example with rstac
stac_endpoint <- "https://catalogue.dataspace.copernicus.eu/stac"
catalog <- stac(stac_endpoint)
products <- stac_search(
  q = catalog,
  collections = "SENTINEL-2",
  datetime = "2016-01-01/2018-01-31"
) |>
  ext_filter(
    tileId == "34UCF" &&
      productType == "S2MSI2A" &&
      id \%like\% "\%\\\\_N0500\\\\_\%"
  ) |>
  post_request() |>
  items_fetch()
}
}
