% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cloud_s2.R
\name{save_cloud_s2}
\alias{save_cloud_s2}
\title{This function saves cloud masks.
'CloudMask_Binary' is default binary mask with 0 where clouds are detected and 1 for clean pixels
'CloudMask_RAW' is the original cloud layer produced by atmospheric correction algorithm
--> may be useful to refine cloud mask}
\usage{
save_cloud_s2(
  S2_stars,
  Cloud_path,
  S2source = "SAFE",
  footprint = NULL,
  SaveRaw = FALSE,
  MaxChunk = 256
)
}
\arguments{
\item{S2_stars}{list. stars object containing raster data. Can be produced with function extract_from_S2_L2A}

\item{Cloud_path}{character.}

\item{S2source}{character.}

\item{footprint}{character. path for vector file defining footprint of interest in the image}

\item{SaveRaw}{boolean. should the original cloud mask layer be saved?}

\item{MaxChunk}{numeric. Size of individual chunks to be written (in Mb)}
}
\value{
list of CloudMasks (binary mask, and raw mask if required)
}
\description{
This function saves cloud masks.
'CloudMask_Binary' is default binary mask with 0 where clouds are detected and 1 for clean pixels
'CloudMask_RAW' is the original cloud layer produced by atmospheric correction algorithm
--> may be useful to refine cloud mask
}
