% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{save_reflectance_s2}
\alias{save_reflectance_s2}
\title{This function saves reflectance files}
\usage{
save_reflectance_s2(
  S2_stars,
  Refl_path,
  Format = "ENVI_BIL",
  datatype = "INT2S",
  S2Sat = NULL,
  tile_S2 = NULL,
  dateAcq_S2 = NULL,
  MTD = NULL
)
}
\arguments{
\item{S2_stars}{list. stars object containing raster data. Can be produced with function extract_from_S2_L2A}

\item{Refl_path}{character. path for reflectance file to be stored}

\item{Format}{character. file format for reflectance data}

\item{datatype}{character. data type (integer, float, 16bits, 32bits...)}

\item{S2Sat}{character. Sentinel-2 mission ('2A' or '2B')}

\item{tile_S2}{character. S2 tile name (2 numbers + 3 letters)}

\item{dateAcq_S2}{double. date of acquisition}

\item{MTD}{character. path for metadata file}
}
\value{
None
}
\description{
This function saves reflectance files
}
