% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_s2_cdse.R
\name{search_items}
\alias{search_items}
\title{search items from Copernicus DataSpace}
\usage{
search_items(
  collection = "SENTINEL-2",
  asset_names = "PRODUCT",
  productType = "S2MSI2A",
  tileId = NULL,
  bbox = NULL,
  SpatVect = NULL,
  cloudcover = 100,
  datetime
)
}
\arguments{
\item{collection}{character. 'SENTINEL-2' as default}

\item{asset_names}{character.}

\item{productType}{character.}

\item{tileId}{character.}

\item{bbox}{numeric.}

\item{SpatVect}{SpatVector.}

\item{cloudcover}{numeric.}

\item{datetime}{character.}
}
\value{
list of items corresponding to input parameters
}
\description{
search items from Copernicus DataSpace
}
