% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options_preprocS2.R
\name{set_options_preprocS2}
\alias{set_options_preprocS2}
\title{set options}
\usage{
set_options_preprocS2(fun, options = NULL)
}
\arguments{
\item{fun}{character. name of the function which has optional parameters}

\item{options}{list. including
\itemize{
\item cloudcover
\item path_S2_tiling_grid
\item overwrite
\item geom_acq
\item additional_process
\item bands_to_correct
\item fraction_vegetation
\item radiometric_filter
}}
}
\value{
options with default values when missing
}
\description{
set options
}
