% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_s2_collection.R
\name{test_s2_collection}
\alias{test_s2_collection}
\title{test S2 collection for a 10 km x 10 km tile of a given aoi}
\usage{
test_s2_collection(
  aoi_path,
  datetime,
  output_dir,
  site_name = NULL,
  stac_info,
  mask_path = NULL,
  cellsize,
  collection
)
}
\arguments{
\item{aoi_path}{character. path for vector file defining aoi}

\item{datetime}{character or date or list defining date of acquisition or time range}

\item{output_dir}{character. path for output directory}

\item{site_name}{character. name of the study site}

\item{stac_info}{list. stac provider, name of collection url for stac}

\item{mask_path}{character.}

\item{cellsize}{numeric.}

\item{collection}{character.}
}
\value{
plots list of plots
}
\description{
test S2 collection for a 10 km x 10 km tile of a given aoi
}
