% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{vectorize_raster_extent}
\alias{vectorize_raster_extent}
\title{save raster footprint as vector file}
\usage{
vectorize_raster_extent(path_raster, path_vector, driver = "ESRI Shapefile")
}
\arguments{
\item{path_raster}{character. path for a raster file}

\item{path_vector}{character. path for a vector file}

\item{driver}{character. driver for vector}
}
\value{
None
}
\description{
save raster footprint as vector file
}
