% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_geomplot.R
\name{write_geomplot}
\alias{write_geomplot}
\title{extracts geometry of acquisition for each plot}
\usage{
write_geomplot(DatesofAcq, geom_dir, geom_subdir, list_aoi, p = NULL)
}
\arguments{
\item{DatesofAcq}{date. date of acquisition}

\item{geom_dir}{character. path for geometry of acquisition main dir}

\item{geom_subdir}{character. path for geometry of acquisition subdir}

\item{list_aoi}{list.}

\item{p}{list.}
}
\value{
list of collections per plot
}
\description{
extracts geometry of acquisition for each plot
}
