% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{write_rasterStack_ENVI}
\alias{write_rasterStack_ENVI}
\title{This function writes a raster Stack object into a ENVI raster file}
\usage{
write_rasterStack_ENVI(
  StackObj,
  StackPath,
  Bands,
  datatype = "INT2S",
  sensor = "Unknown",
  Stretch = FALSE
)
}
\arguments{
\item{StackObj}{list. raster stack}

\item{StackPath}{character. path where to store the stack}

\item{Bands}{list. should include 'bandname', and if possible 'wavelength'}

\item{datatype}{character. should be INT2S or FLT4S for example}

\item{sensor}{character. Name of the sensor used to acquire the image}

\item{Stretch}{boolean. Set TRUE to get 10\% stretching at display for reflectance, mentioned in hdr only}
}
\value{
None
}
\description{
This function writes a raster Stack object into a ENVI raster file
}
