% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{write_starsStack_ENVI}
\alias{write_starsStack_ENVI}
\title{This function writes a stars Stack object into a ENVI raster file}
\usage{
write_starsStack_ENVI(
  StarsObj,
  dsn,
  datatype = "Int16",
  Bands,
  sensor = "Unknown",
  Stretch = FALSE
)
}
\arguments{
\item{StarsObj}{list. stars stack (read with along = 'band')}

\item{dsn}{character. path where to store the stack}

\item{datatype}{character. should be Int16 or Float64 for example}

\item{Bands}{list. should include 'bandname', and if possible 'wavelength'}

\item{sensor}{character. Name of the sensor used to acquire the image}

\item{Stretch}{boolean. Set TRUE to get 10\% stretching at display for reflectance, mentioned in hdr only}
}
\value{
None
}
\description{
This function writes a stars Stack object into a ENVI raster file
}
