% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{write_template_MSIL2A}
\alias{write_template_MSIL2A}
\title{This function is made to correctly create the MTD_MSIL2A.xml for each dnowloaded date using the MTD_MSIL2A.xml template
programmed by romain.dous@inrae.fr}
\usage{
write_template_MSIL2A(features_Acq, AcqDir, Acq)
}
\arguments{
\item{features_Acq}{list. list of features datas link to the studied acquisition}

\item{AcqDir}{character. Path for the acquisition}

\item{Acq}{character. data of the acquisiton}
}
\description{
This function is made to correctly create the MTD_MSIL2A.xml for each dnowloaded date using the MTD_MSIL2A.xml template
programmed by romain.dous@inrae.fr
}
