% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{write_template_TL}
\alias{write_template_TL}
\title{This function is made to correctly create the MTD_TL.xml for each dnowloaded date using the MTD_TL.xml template
programmed by romain.dous@inrae.fr}
\usage{
write_template_TL(features_Acq, AcqDir)
}
\arguments{
\item{features_Acq}{list. list of features datas link to the studied acquisition}

\item{AcqDir}{character. Path for the acquisition}
}
\description{
This function is made to correctly create the MTD_TL.xml for each dnowloaded date using the MTD_TL.xml template
programmed by romain.dous@inrae.fr
}
