% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_radiometric_filter.R
\name{apply_radiometric_filter}
\alias{apply_radiometric_filter}
\title{updates vegetation mask based on initial SCL and readiometric thresholds}
\usage{
apply_radiometric_filter(
  S2_item,
  acq,
  iChar,
  raster_dir,
  cloudmask,
  mask_path = NULL,
  aoiplot,
  fraction_vegetation,
  siteName = NULL,
  RadiometricFilter = NULL,
  asset_cloud,
  original_clouds = TRUE,
  overwrite = T,
  writeoutput = T
)
}
\arguments{
\item{S2_item}{path for spectral bands required}

\item{acq}{date of acquisition}

\item{iChar}{plot ID}

\item{raster_dir}{directory where rasters are stored}

\item{cloudmask}{spatial raster}

\item{mask_path}{path for binary mask}

\item{aoiplot}{sf.}

\item{fraction_vegetation}{minimum fraction of vegetation required to keep the acquisition}

\item{siteName}{character. name of the study site}

\item{RadiometricFilter}{list. criterions for cloud, shade, non vegetation}

\item{asset_cloud}{character. asset name for cloud mask}

\item{original_clouds}{boolean}

\item{overwrite}{boolean}

\item{writeoutput}{boolean. should output file be saved?}
}
\value{
list of updated masks and valid dates of acquisition
}
\description{
updates vegetation mask based on initial SCL and readiometric thresholds
}
