% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_s2_tiling_grid.R
\name{check_s2_tiling_grid}
\alias{check_s2_tiling_grid}
\title{check if S2 tiling grid is available}
\usage{
check_s2_tiling_grid(path_S2_tiling_grid = NULL)
}
\arguments{
\item{path_S2_tiling_grid}{character. path for S2 tiling grid}
}
\value{
valid path for S2 tiling grid
}
\description{
check if S2 tiling grid is available
}
