% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_geom.R
\name{correct_geom}
\alias{correct_geom}
\title{corrects geometry of acquisition}
\usage{
correct_geom(S2_rast, output_dir, aoi, acq)
}
\arguments{
\item{S2_rast}{S2 spatRaster}

\item{output_dir}{character.}

\item{aoi}{sf object}

\item{acq}{date of acquisition}
}
\value{
list of collections per plot
}
\description{
corrects geometry of acquisition
}
