% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cloudmask.R
\name{download_cloudmask}
\alias{download_cloudmask}
\title{downloads SCL from Planetary}
\usage{
download_cloudmask(
  aoi,
  raster_dir,
  item_collection,
  iChar,
  stac_info,
  asset_names,
  resolution,
  siteName = NULL,
  crs_target = NULL
)
}
\arguments{
\item{aoi}{geometry corresponding to one point}

\item{raster_dir}{directory where rasters are stored}

\item{item_collection}{description of collection to download}

\item{iChar}{plot ID}

\item{stac_info}{list.}

\item{asset_names}{character.}

\item{resolution}{numeric. spatial resolution (10 or 20)}

\item{siteName}{character. name of the study site}

\item{crs_target}{numeric.}
}
\value{
list of collections per plot
}
\description{
downloads SCL from Planetary
}
