% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_s2.R
\name{download_s2}
\alias{download_s2}
\title{downloads S2 reflectance data}
\usage{
download_s2(
  aoi,
  raster_dir,
  collection_path,
  iChar,
  resolution,
  offset = 1000,
  offset_B2 = F,
  collection = "sentinel-2-l2a",
  corr_BRF = F,
  clean_bands = T
)
}
\arguments{
\item{aoi}{geometry corresponding to one point}

\item{raster_dir}{directory where rasters are stored}

\item{collection_path}{path for collection to download}

\item{iChar}{plot ID}

\item{resolution}{numeric. spatial resolution (10 or 20)}

\item{offset}{numeric. S2 offset value}

\item{offset_B2}{boolean.}

\item{collection}{character.}

\item{corr_BRF}{boolean.}

\item{clean_bands}{boolean.}
}
\value{
list of collections per plot
}
\description{
downloads S2 reflectance data
}
