% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_doublons_dateAcq.R
\name{eliminate_doublons_dateAcq}
\alias{eliminate_doublons_dateAcq}
\title{eliminates doublons in terms of date of acquisition from collection}
\usage{
eliminate_doublons_dateAcq(item_collection)
}
\arguments{
\item{item_collection}{collection STAC item collection}
}
\value{
collection STAC item collection
}
\description{
eliminates doublons in terms of date of acquisition from collection
}
