% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_B248_filter.R
\name{get_B248_filter}
\alias{get_B248_filter}
\title{downloads S2 reflectance to update the vegetation mask}
\usage{
get_B248_filter(
  raster_dir,
  mask_path,
  item_collection,
  cloudmasks,
  iChar,
  aoi,
  resolution,
  collection = "sentinel-2-l2a",
  fraction_vegetation,
  offset = 1000,
  RadiometricFilter = NULL,
  siteName,
  crs_target = NULL,
  original_clouds = original_clouds,
  overwrite = F,
  S2_items = NULL,
  writeoutput = T
)
}
\arguments{
\item{raster_dir}{directory where rasters are stored}

\item{mask_path}{directory where rasters are stored}

\item{item_collection}{path for collection to download}

\item{cloudmasks}{spatial raster}

\item{iChar}{plot ID}

\item{aoi}{geometry corresponding to one point}

\item{resolution}{numeric. spatial resolution (10 or 20)}

\item{collection}{character.}

\item{fraction_vegetation}{numeric. minimum \% of vegetatiomn to have on plot footprint}

\item{offset}{numeric. value of S2 offset}

\item{RadiometricFilter}{list. list of radiometric filters for shade, clouds and vegetation}

\item{siteName}{character. name of the study site}

\item{crs_target}{numeric.}

\item{original_clouds}{boolean}

\item{overwrite}{boolean}

\item{S2_items}{list}

\item{writeoutput}{boolean. should output file be saved?}
}
\value{
list of collections per plot
}
\description{
downloads S2 reflectance to update the vegetation mask
}
