% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cloud_asset.R
\name{get_cloud_asset}
\alias{get_cloud_asset}
\title{get name for cloud asset in item collection}
\usage{
get_cloud_asset(stac_info)
}
\arguments{
\item{stac_info}{list.}
}
\value{
asset_names corresponding to cloud or scene classification
}
\description{
get name for cloud asset in item collection
}
