% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collections.R
\name{get_collections}
\alias{get_collections}
\title{gets collections for plot network}
\usage{
get_collections(
  list_aoi,
  s2_tiles = NULL,
  datetime,
  output_dir,
  overwrite = T,
  cloudcover = 100,
  stac_info,
  nbCPU = 1,
  doublecheckColl = T
)
}
\arguments{
\item{list_aoi}{list. list of simple features}

\item{s2_tiles}{list. list of S2 tiles including each sf from list_aoi}

\item{datetime}{list. define period 'from' / 'to' day of acquisition}

\item{output_dir}{character. path for output directory}

\item{overwrite}{boolean. should collection and S2 data be overwritten?}

\item{cloudcover}{numeric. maximum cloud cover over tile}

\item{stac_info}{list. stac provider, name of collection url for stac}

\item{nbCPU}{numeric. number of CPU}

\item{doublecheckColl}{boolean. set TRUE to double check collection}
}
\value{
list of collections per plot
}
\description{
gets collections for plot network
}
