% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dateAcq.R
\name{get_dateAcq}
\alias{get_dateAcq}
\title{gets acquisition date from planetary URL}
\usage{
get_dateAcq(band_url = NULL, S2product = NULL)
}
\arguments{
\item{band_url}{character.}

\item{S2product}{character.}
}
\value{
DateAcq dates of acquisition
}
\description{
gets acquisition date from planetary URL
}
