% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_aoi.R
\name{get_grid_aoi}
\alias{get_grid_aoi}
\title{return vector grid corresponding to an area of interest}
\usage{
get_grid_aoi(
  aoi_path,
  cellsize = 10000,
  crs_target = 4326,
  dsn_grid = NULL,
  dsn_centroid = NULL
)
}
\arguments{
\item{aoi_path}{character. path for vector file defining aoi}

\item{cellsize}{numeric. square cell size in meters}

\item{crs_target}{numeric. CRS corresponding to the output grid cell}

\item{dsn_grid}{character. path corresponding to grid}

\item{dsn_centroid}{character. path corresponding to centroids of grid}
}
\value{
dsn_grid path for grid of aoi
}
\description{
return vector grid corresponding to an area of interest
}
