% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_plots.R
\name{get_missing_plots}
\alias{get_missing_plots}
\title{return vector grid corresponding to an area of interest}
\usage{
get_missing_plots(plots, pursue_existing, datetime, output_dir, pattern = NULL)
}
\arguments{
\item{plots}{character. list of vector plots}

\item{pursue_existing}{boolean. set TRUE if continue already started process}

\item{datetime}{list. datetime of acquisition}

\item{output_dir}{character. output dir}

\item{pattern}{character. pattern to look for in output_dir}
}
\value{
missing plot ID
}
\description{
return vector grid corresponding to an area of interest
}
