% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s2_collection.R
\name{get_s2_collection}
\alias{get_s2_collection}
\title{downloads Sentinel-2 time series for a vectorized plot network}
\usage{
get_s2_collection(
  plots,
  S2tiles = NULL,
  datetime,
  output_dir,
  mask_path = NULL,
  cloudcover = 100,
  fraction_vegetation = 5,
  resolution = 10,
  collection = "sentinel-2-l2a",
  stac_url = NULL,
  overwrite = F,
  nbCPU = 1,
  doublecheckColl = T,
  offset = 1000,
  offset_B2 = F,
  corr_BRF = F,
  RadiometricFilter = NULL,
  siteName = NULL,
  rast_out = T,
  additional_process = NULL,
  crs_target = NULL,
  original_clouds = TRUE,
  argsin = NULL,
  writeoutput = T,
  bands2correct = c("B8A", "B11", "B12")
)
}
\arguments{
\item{plots}{list. sf objects corresponding to polygons}

\item{S2tiles}{list. S2 tiles corresponding to the plots}

\item{datetime}{list. define period 'from' / 'to' day of acquisition}

\item{output_dir}{character. output directory}

\item{mask_path}{character. path for mask}

\item{cloudcover}{numeric. maximum cloud cover over tile}

\item{fraction_vegetation}{numeric. minimum fraction vegetation to consider acquisition}

\item{resolution}{numeric. spatial resolution (10 or 20)}

\item{collection}{character. collection}

\item{stac_url}{character. STAC endpoint}

\item{overwrite}{boolean. should collection and S2 data be overwritten?}

\item{nbCPU}{numeric. number of threads to work with}

\item{doublecheckColl}{boolean.}

\item{offset}{numeric. S2 reflectance offset}

\item{offset_B2}{boolean. should an offset be applied to normalize B2?}

\item{corr_BRF}{boolean.}

\item{RadiometricFilter}{list.}

\item{siteName}{character. name of the study site}

\item{rast_out}{boolean. should S2 SpatRaster be obtained as output?}

\item{additional_process}{additional process to be applied to S2_items once downloaded}

\item{crs_target}{numeric.}

\item{original_clouds}{should original cloud mask be used or not?}

\item{argsin}{list}

\item{writeoutput}{boolean. should output file be saved?}

\item{bands2correct}{character. name of bands to correct from geometry}
}
\value{
list of collections per plot
}
\description{
downloads Sentinel-2 time series for a vectorized plot network
}
