% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s2_raster.R
\name{get_s2_raster}
\alias{get_s2_raster}
\title{return S2 tiles intersecting with the footprint of a vector file}
\usage{
get_s2_raster(
  aoi_path = NULL,
  bbox = NULL,
  datetime,
  output_dir,
  cloudcover = 100,
  site = NULL,
  path_S2tilinggrid = NULL,
  overwrite = T,
  geomAcq = F,
  authentication = NULL,
  collection = "sentinel-2-l2a",
  stac_url = NULL
)
}
\arguments{
\item{aoi_path}{character. path for vector file defining aoi}

\item{bbox}{bbox}

\item{datetime}{character or date or list defining date of acquisition or time range}

\item{output_dir}{character. path for output directory}

\item{cloudcover}{numeric. cloud cover}

\item{site}{character. name of the study site}

\item{path_S2tilinggrid}{character. path for the Sentinel-2_tiling_grid.kml file}

\item{overwrite}{boolean.}

\item{geomAcq}{boolean.}

\item{authentication}{list. authentication to CDSE}

\item{collection}{character.}

\item{stac_url}{character.}
}
\value{
S2tiles list of tiles corresponding to plots
}
\description{
return S2 tiles intersecting with the footprint of a vector file
}
