% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s2_tiling.R
\name{get_s2_tiling}
\alias{get_s2_tiling}
\title{return S2 tiles intersecting with the footprint of a vector file}
\usage{
get_s2_tiling(
  plots,
  aoi_path,
  datetime,
  output_dir,
  cloudcover = 100,
  site = NULL,
  path_S2tilinggrid = NULL,
  overwrite = T,
  collection = "sentinel-2-l2a",
  geomAcq = F,
  nbCPU = 1,
  authentication = NULL,
  mask_path = NULL,
  resolution = 10,
  fraction_vegetation = 5,
  stac_url = NULL,
  doublecheckColl = T,
  offset = 1000,
  offset_B2 = F,
  corr_BRF = F,
  RadiometricFilter = NULL,
  clean_bands = T
)
}
\arguments{
\item{plots}{list.}

\item{aoi_path}{character. path for vector file defining aoi}

\item{datetime}{character or date or list defining date of acquisition or time range}

\item{output_dir}{character. path for output directory}

\item{cloudcover}{numeric.}

\item{site}{character. name of the study site}

\item{path_S2tilinggrid}{character. path for the Sentinel-2_tiling_grid.kml file}

\item{overwrite}{boolean.}

\item{collection}{character.}

\item{geomAcq}{boolean.}

\item{nbCPU}{numeric.}

\item{authentication}{list.}

\item{mask_path}{character.}

\item{resolution}{numeric.}

\item{fraction_vegetation}{numeric.}

\item{stac_url}{character.}

\item{doublecheckColl}{boolean.}

\item{offset}{numeric.}

\item{offset_B2}{boolean.}

\item{corr_BRF}{boolean.}

\item{RadiometricFilter}{list.}

\item{clean_bands}{boolean}
}
\value{
S2tiles list of tiles corresponding to plots
}
\description{
return S2 tiles intersecting with the footprint of a vector file
}
