% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{mosaic_rasters}
\alias{mosaic_rasters}
\title{mosaicing a set of rasters}
\usage{
mosaic_rasters(list_rasters, dst_mosaic, Stretch = FALSE)
}
\arguments{
\item{list_rasters}{character. list of paths corresponding to rasters to mosaic}

\item{dst_mosaic}{character. path and name of mosaic produced}

\item{Stretch}{boolean. Set TRUE to get 10\% stretching at display for reflectance, mentioned in hdr only}
}
\value{
None
}
\description{
mosaicing a set of rasters
}
