% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_preprocess_S2.R
\name{write_Stack_S2}
\alias{write_Stack_S2}
\title{This function writes a stars object into a raster file}
\usage{
write_Stack_S2(
  Stars_S2,
  Stars_Spectral,
  Refl_path,
  Format = "ENVI_BIL",
  datatype = "INT2S",
  sensor = "Unknown"
)
}
\arguments{
\item{Stars_S2}{list. stars object containing raster data. Can be produced with function Crop_n_resample_S2}

\item{Stars_Spectral}{list. band name to be saved in the stack and spectral bands corresponding to the image}

\item{Refl_path}{character. path where to store the image}

\item{Format}{character. default = ENVI BIL. otheerwise use gdal drivers}

\item{datatype}{character. should be INT2S or FLT4S for example}

\item{sensor}{character. Name of the sensor used to acquire the image}
}
\value{
None
}
\description{
This function writes a stars object into a raster file
}
